Add a check to verify that the driver has been initialised within Browser#deleteCookies and Browser#refresh

+++ src/main/java/org/dvsa/testing/lib/browser/Browser.java
@@ -198,11 +198,13 @@ public class Browser {
     }
 
     public static void deleteCookies() {
-        getDriver().manage().deleteAllCookies();
+        if (!isUninitialisedOrClosed())
+            getDriver().manage().deleteAllCookies();
     }
 
     public static void refresh() {
-        getDriver().navigate().refresh();
+        if (!isUninitialisedOrClosed())
+            getDriver().navigate().refresh();
     }
 
 }
